separate logics from useOthello
